requirements.txt
africastalking==1.2.7
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==4.2.13
djangorestframework==3.15.2
idna==3.7
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
schema==0.7.7
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
